chore(cloudflare): rename unified entry to fetch handler
