+++ apps/web/worker/index.ts
@@ -1,5 +1,5 @@
 /** Cloudflare Worker entry point for web-specific APIs and scheduled maintenance. */
-import handler from "vinext/server/entry";
+import handler from "vinext/server/fetch-handler";
 
 type Env = {
   ASSETS: Fetcher;
