+++ examples/nextra-docs-template/wrangler.jsonc
@@ -3,7 +3,7 @@
   "name": "nextra-docs-template",
   "compatibility_date": "2026-02-12",
   "compatibility_flags": ["nodejs_compat"],
-  "main": "vinext/server/entry",
+  "main": "vinext/server/fetch-handler",
   "preview_urls": true,
   "assets": {
     "not_found_handling": "none"
