+++ examples/pages-router-cloudflare/wrangler.jsonc
@@ -3,7 +3,7 @@
   "name": "pages-router-cloudflare",
   "compatibility_date": "2026-02-12",
   "compatibility_flags": ["nodejs_compat"],
-  "main": "vinext/server/entry",
+  "main": "vinext/server/fetch-handler",
   "preview_urls": true,
   "assets": {
     "directory": "dist/client",
