+++ examples/realworld-api-rest/wrangler.jsonc
@@ -3,7 +3,7 @@
   "name": "realworld-api-rest",
   "compatibility_date": "2026-02-12",
   "compatibility_flags": ["nodejs_compat"],
-  "main": "vinext/server/entry",
+  "main": "vinext/server/fetch-handler",
   "preview_urls": true,
   "assets": {
     "directory": "dist/client",
