+++ tests/deploy.test.ts
@@ -195,7 +195,7 @@ describe("deploy environment validation", () => {
     writeFile(
       tmpDir,
       "wrangler.jsonc",
-      '{"main":"vinext/server/entry","assets":{"directory":"dist/client"}}\n',
+      '{"main":"vinext/server/fetch-handler","assets":{"directory":"dist/client"}}\n',
     );
 
     await expect(deploy({ root: tmpDir, dryRun: true })).rejects.not.toThrow("Worker entry");
@@ -637,7 +637,7 @@ describe("generateWranglerConfig", () => {
 
     expect(parsed.name).toBe(info.projectName);
     expect(parsed.compatibility_flags).toContain("nodejs_compat");
-    expect(parsed.main).toBe("vinext/server/entry");
+    expect(parsed.main).toBe("vinext/server/fetch-handler");
     expect(parsed.assets).toEqual({
       directory: "dist/client",
       not_found_handling: "none",
@@ -646,14 +646,14 @@ describe("generateWranglerConfig", () => {
     expect(parsed.$schema).toBe("node_modules/wrangler/config-schema.json");
   });
 
-  it("points Pages Router apps at the built-in worker entry", () => {
+  it("points Pages Router apps at the built-in fetch handler", () => {
     mkdir(tmpDir, "pages");
     writeFile(tmpDir, "pages/index.tsx", "export default function Page() { return null; }");
     const info = detectProject(tmpDir);
     const config = generateWranglerConfig(info);
     const parsed = JSON.parse(config);
 
-    expect(parsed.main).toBe("vinext/server/entry");
+    expect(parsed.main).toBe("vinext/server/fetch-handler");
   });
 
   it("sets compatibility_date to today", () => {
@@ -1084,9 +1084,9 @@ describe("generatePagesRouterWorkerEntry", () => {
     expect(content).toContain("env.ASSETS!.fetch(assetRequest)");
   });
 
-  it("exports the built-in worker entries", () => {
+  it("exports the built-in fetch handler and router-specific worker entries", () => {
     const exportsMap = readVinextPackageExports();
-    expect(hasPackageExport(exportsMap, "./server/entry")).toBe(true);
+    expect(hasPackageExport(exportsMap, "./server/fetch-handler")).toBe(true);
     expect(hasPackageExport(exportsMap, "./server/app-router-entry")).toBe(true);
     expect(hasPackageExport(exportsMap, "./server/pages-router-entry")).toBe(true);
   });
