+++ tests/e2e/cloudflare-workers/fixture/worker/index.mjs
@@ -1,4 +1,4 @@
-import handler from "vinext/server/entry";
+import handler from "vinext/server/fetch-handler";
 
 const CSP = "script-src 'nonce-vinext-test-nonce' 'strict-dynamic';";
