+++ tests/fixtures/cf-app-basic/wrangler.jsonc
@@ -2,7 +2,7 @@
   "name": "cf-app-basic",
   "compatibility_date": "2026-02-12",
   "compatibility_flags": ["nodejs_compat"],
-  "main": "vinext/server/entry",
+  "main": "vinext/server/fetch-handler",
   "assets": {
     "not_found_handling": "none",
     "binding": "ASSETS",
