+++ tests/fixtures/ppr-impact-demo/wrangler.jsonc
@@ -3,5 +3,5 @@
   "name": "ppr-impact-demo",
   "compatibility_date": "2026-04-08",
   "compatibility_flags": ["nodejs_compat"],
-  "main": "vinext/server/entry",
+  "main": "vinext/server/fetch-handler",
 }
