+++ tests/init.test.ts
@@ -483,11 +483,11 @@ describe("init — basic functionality", () => {
     expect(readFile(tmpDir, "vite.config.ts")).not.toContain("cdn:");
     expect(fs.existsSync(path.join(tmpDir, "worker", "index.ts"))).toBe(false);
     expect(JSON.parse(readFile(tmpDir, "wrangler.jsonc"))).toMatchObject({
-      main: "vinext/server/entry",
+      main: "vinext/server/fetch-handler",
     });
   });
 
-  it("uses the built-in worker entry for Pages Router Cloudflare init", async () => {
+  it("uses the built-in fetch handler for Pages Router Cloudflare init", async () => {
     setupProject(tmpDir, { router: "pages" });
 
     const { result } = await runInit(tmpDir, { platform: "cloudflare" });
@@ -495,7 +495,7 @@ describe("init — basic functionality", () => {
     expect(result.generatedPlatformFiles).toEqual(["wrangler.jsonc"]);
     expect(fs.existsSync(path.join(tmpDir, "worker", "index.ts"))).toBe(false);
     expect(JSON.parse(readFile(tmpDir, "wrangler.jsonc"))).toMatchObject({
-      main: "vinext/server/entry",
+      main: "vinext/server/fetch-handler",
     });
   });
 
@@ -600,7 +600,7 @@ describe("init — basic functionality", () => {
 
     expect(result.generatedPlatformFiles).toContain("wrangler.jsonc");
     expect(JSON.parse(readFile(tmpDir, "wrangler.jsonc"))).toMatchObject({
-      main: "vinext/server/entry",
+      main: "vinext/server/fetch-handler",
     });
     expect(readFile(tmpDir, "cloudflare.config.ts")).toBe("export default {};\n");
   });
